ajustes centros de entrenamiento

+++ src/app/training-centers/training-centers.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { MatPaginator, PageEvent } from '@angular/material/paginator';
 import { MatTableDataSource } from '@angular/material/table';
+import { ConfirmDialogComponent } from '../confirm-dialog/confirm-dialog.component';
 import { TrainingCenterListDto } from '../intefaces/training-center-list-dto';
 import { TrainingCenterService } from '../services/training-center.service';
 
@@ -10,15 +12,19 @@ import { TrainingCenterService } from '../services/training-center.service';
   styleUrls: ['./training-centers.component.css']
 })
 export class TrainingCentersComponent implements OnInit {
+  confirmed = false;
   countRegisters: number = 0
   initPageSize: number = 5
-  displayedColumns: string[] = ['Code', 'Name', 'Update','actions'];
+  displayedColumns: string[] = ['Code', 'Name','actions'];
   ELEMENT_DATA: TrainingCenterListDto[] = [];
   dataSource = new MatTableDataSource<TrainingCenterListDto>(this.ELEMENT_DATA);
 
   @ViewChild(MatPaginator) paginator! : MatPaginator;
 
-  constructor( private traningCenterService : TrainingCenterService) { }
+  constructor( 
+    private traningCenterService : TrainingCenterService,
+    private dialog: MatDialog
+    ) { }
 
   ngOnInit(): void {
     this.traningCenterService.getAllTrainingCenter(0,this.initPageSize,true).subscribe(
@@ -44,15 +50,25 @@ export class TrainingCentersComponent implements OnInit {
         this.countRegisters = data["totalDbRegistros"]
       })*/
   }
+
   onDelete(Id: string) {
-    this.traningCenterService.DeleteByIdTraningCenter(Id).subscribe(response => 
-      this.traningCenterService.getAllTrainingCenter(0,this.initPageSize,true).subscribe(
-        data =>
-        {
-          this.ELEMENT_DATA = data["registros"];  
-          this.countRegisters = data["totalDbRegistros"];
-        }
-      ))
+    const dialogRef = this.dialog.open(ConfirmDialogComponent, {
+      data: {type: 'alert-red',title: '¿Está seguro que desea eliminar este registro?', message: 'Esta operación es irreversible', textButton: 'Eliminar' }
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      this.confirmed = result;
+      if (this.confirmed) {
+        this.traningCenterService.DeleteByIdTraningCenter(Id).subscribe(response => 
+              this.traningCenterService.getAllTrainingCenter(0,this.initPageSize,true).subscribe(
+                data =>
+                {
+                  this.ELEMENT_DATA = data["registros"];  
+                  this.countRegisters = data["totalDbRegistros"];
+                }
+              ))
+      }
+    });
   }
 }
